Squashed commit of the following:
commit b666413acc0eacbd0d04e8e278da47d8cf01ec16
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Thu Oct 26 21:35:12 2023 -0300

    Squashed commit of the following:

    commit 812f0563f205e9e951a15c7eda885fe02767e7cb
    Author: Manuel Sturla <manuelsturla@gmail.com>
    Date:   Thu Oct 26 20:47:10 2023 -0300

        fix test parties view

    commit 99571abeed194e1cbd74d0d66e4044796e099a84
    Author: Manuel Sturla <manuelsturla@gmail.com>
    Date:   Thu Oct 26 20:29:50 2023 -0300

        Parties View refactor [WIP]

    commit ead0e43c6fb50ae99543c730be5436b908de6e1f
    Author: Manuel Sturla <manuelsturla@gmail.com>
    Date:   Sat Oct 21 10:25:50 2023 -0300

        add search and ordering of party view

    commit 26159da0afe631e4674b634de84036b7149cdedd
    Merge: dd67205 7e64745
    Author: Manuel Sturla <manuelsturla@gmail.com>
    Date:   Mon Oct 16 20:41:03 2023 -0300

        Merge pull request #2 from VotaInformado/refactor-votes-summary

        Add votes necessary endpoints

commit bff582e477d5657268ccf5c48e1db4eb4e3a675a
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Thu Oct 26 21:34:03 2023 -0300

    adds test for filtered data retrieval

commit f3f4956bf22903b332c523d1cb36fc73e560e51f
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Tue Oct 24 21:17:07 2023 -0300

    adds law projects endpoint for neural network

commit c3798ba7ade3365173fe651fa471a75a594c50fe
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Tue Oct 24 20:34:20 2023 -0300

    completes last commit

commit 83dec6558a746354642a23fb2bf66e182b951f5b
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Tue Oct 24 20:34:06 2023 -0300

    adds votes view for neural network data

commit 22b7eb66a47b2b4dd15cda83ab3046c35fccbe96
Author: Andrés Fernández <anfernandez@fi.uba.ar>
Date:   Mon Oct 23 22:48:59 2023 -0300

    adds votes endpoint for neural_network microservice

+++ recoleccion/management/commands/create_sub_party.py
@@ -1,4 +1,5 @@
 import pandas as pd
+from django.db import IntegrityError
 
 # Base command
 from django.core.management.base import BaseCommand
@@ -7,9 +8,7 @@ from django.core.management.base import BaseCommand
 from django.db import transaction
 
 # Components
-from recoleccion.components.data_sources.day_orders_source import DeputiesDayOrderSource
-from recoleccion.components.writers.law_projects_writer import LawProjectsWriter
-from recoleccion.models import DeputySeat, Party, PartyDenomination, PartyRelationTypes, SenateSeat
+from recoleccion.models import Authorship, DeputySeat, Party, PartyDenomination, PartyRelationTypes, SenateSeat, Vote
 from recoleccion.utils.custom_logger import CustomLogger
 
 
@@ -27,17 +26,31 @@ class Command(BaseCommand):
         sub_party_id = options["sub_party_id"]
         relation_type = options["relation_type"]
         party = Party.objects.get(id=party_id)
-        sub_party = Party.objects.get(id=sub_party_id)
         self.logger.info(f"Party: {party.main_denomination}")
-        self.logger.info(f"Sub party: {sub_party.main_denomination}")
+        sub_party = Party.objects.filter(id=sub_party_id).first()
+        if not sub_party:
+            self.logger.info(f"Sub party {sub_party_id} has already been deleted...")
+        else:
+            self.logger.info(f"Sub party: {sub_party.main_denomination}")
         sub_party_senate_seats = SenateSeat.objects.filter(party_id=sub_party_id)
         sub_party_deputy_seats = DeputySeat.objects.filter(party_id=sub_party_id)
+        sub_party_votes = Vote.objects.filter(party_id=sub_party_id)
+        sub_party_autorships = Authorship.objects.filter(party_id=sub_party_id)
         self.logger.info(f"Updated {len(sub_party_senate_seats)} senate seats")
         sub_party_senate_seats.update(party_id=party_id)
         self.logger.info(f"Updated {len(sub_party_deputy_seats)} deputy seats")
         sub_party_deputy_seats.update(party_id=party_id)
-        PartyDenomination.objects.create(
-            party=party, denomination=sub_party.main_denomination, relation_type=relation_type
-        )
+        self.logger.info(f"Updated {len(sub_party_votes)} votes")
+        sub_party_votes.update(party_id=party_id)
+        self.logger.info(f"Updated {len(sub_party_autorships)} authorships")
+        sub_party_autorships.update(party_id=party_id)
+        if not sub_party:
+            return
         sub_party.delete()
+        try:
+            PartyDenomination.objects.create(
+                party=party, denomination=sub_party.main_denomination, relation_type=relation_type
+            )
+        except IntegrityError:
+            self.logger.info(f"Sub party {sub_party.main_denomination} already exists")
         self.logger.info(f"Deleted sub party {sub_party.main_denomination}")
